src/use-cases/get-user-profile.spec.ts
import { beforeEach, describe, expect, it } from 'vitest'
import { InMemoryUsersRepository } from '@/repositories/in-memory/in-memory-users-repositorys'
import { hash } from 'bcryptjs'
import { GetUserProfileCase } from './get-user-profile'
import { ResourceNotFoundError } from './errors/resourcec-not-found-error'

let usersRepository: InMemoryUsersRepository
let sut: GetUserProfileCase

describe('Get User Profile Use Case', () => {
  beforeEach(() => {
    usersRepository = new InMemoryUsersRepository()
    sut = new GetUserProfileCase(usersRepository)
  })

  it('should be able to get user profile', async () => {
    const createdUser = await usersRepository.create({
      name: 'flavio leliz',
      email: 'example@email.com',
      password_hash: await hash('123', 6),
    })

    const { user } = await sut.execute({
      userId: createdUser.id,
    })

    expect(user.id).toEqual(expect.any(String))
    expect(user.name).toEqual('flavio leliz')
  })

  it('should not be able to get user profile with wrong id', async () => {
    await expect(() =>
      sut.execute({
        userId: 'non-exist-id',
      }),
    ).rejects.toBeInstanceOf(ResourceNotFoundError)
  })
})
